requirements.txt
# Core dependencies for AI Agent Company System
python-dotenv>=1.0.0
requests>=2.31.0
langchain-groq>=0.1.0
groq>=0.4.0

# Database dependencies for centralized memory system
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
supabase>=2.0.0

# Memory and AI
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API Gateway and Web Server (New Features)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pyjwt>=2.8.0
pydantic>=2.5.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# External API Integrations (New Departments)
# Marketing - Google Workspace
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0

# Finance - Payment APIs
stripe>=7.0.0
razorpay>=1.3.0
plaid-python>=9.0.0

# Engineering - GitHub API
PyGithub>=1.59.0

# Support - Zendesk API (optional)
zenpy>=2.0.25

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0# Doc
ument generation dependencies (New Feature)
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.0.10
reportlab>=3.6.0

# Additional security and file management
bcrypt>=3.2.0
cryptography>=3.4.8
Pillow>=9.0.0